Altered several small interactions to ensure that this bot plays well with other bots.
Added a one-second delay to ensure that there's no ambiguity in response time.
Altered response logic to permit responding to bots (infinite loops may still be possible if another bot responds to this bot with [bracketed text]).
Altered reaction logic so that members of Discord can add reactions for members of Gemp if they produce a menu.